Split out the elaborator from `CloneMap`.
This opens the possibility to doing something other than literally
emitting a sequence of `clone`s, but some more cleanup would first be
required to make this truly viable.
