+++ creusot/src/backend/clone_map.rs
@@ -1,14 +1,8 @@
 #![allow(deprecated)]
 
-use std::{collections::HashSet, fmt::Display, iter};
-
 use heck::ToUpperCamelCase;
 use indexmap::{IndexMap, IndexSet};
-use itertools::Itertools;
-use petgraph::{
-    algo::is_cyclic_directed, graphmap::DiGraphMap, visit::DfsPostOrder, Direction,
-    EdgeDirection::Outgoing,
-};
+use petgraph::{graphmap::DiGraphMap, visit::DfsPostOrder, EdgeDirection::Outgoing};
 use rustc_hir::{
     def::{DefKind, Namespace},
     def_id::DefId,
@@ -17,20 +11,29 @@ use rustc_middle::ty::{
     self, subst::SubstsRef, AliasKind, AliasTy, ParamEnv, Ty, TyCtxt, TyKind, TypeFoldable,
     TypeSuperVisitable, TypeVisitor,
 };
-use rustc_span::{Symbol, DUMMY_SP};
+use rustc_span::Symbol;
 use rustc_target::abi::FieldIdx;
 
 use why3::{
-    declaration::{CloneKind, CloneSubst, Decl, DeclClone, Use},
+    declaration::{CloneKind, CloneSubst, Decl, Use},
     Ident, QName,
 };
 
 use crate::{
-    backend::{self, dependency::Dependency, interface},
+    backend::{clone_map::elaborator::CloneElaborator, dependency::Dependency, interface},
     ctx::*,
     translation::traits,
-    util::{self, get_builtin, ident_of, ident_of_ty, inv_module_name, item_name, module_name},
+    util::{self, ident_of, ident_of_ty, inv_module_name, item_name, module_name},
 };
+use rustc_macros::{TyDecodable, TyEncodable};
+
+use super::{
+    ty_inv::{self, TyInvKind},
+    TransId, Why3Generator,
+};
+
+
+mod elaborator;
 
 // Prelude modules
 #[allow(dead_code)]
@@ -216,9 +219,6 @@ pub struct CloneMap<'tcx> {
 
     // Internal state to determine whether clones should be public or not
     dep_level: CloneLevel,
-
-    // Used to ensure we only have a single `use` per type.
-    used_types: IndexSet<DefId>,
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, TyEncodable, TyDecodable, Hash)]
@@ -246,13 +246,6 @@ impl Into<CloneKind> for Kind {
     }
 }
 
-use rustc_macros::{TyDecodable, TyEncodable};
-
-use super::{
-    ty::ty_param_names,
-    ty_inv::{self, TyInvKind},
-    TransId, Why3Generator,
-};
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, TyEncodable, TyDecodable)]
 enum CloneOpacity {
@@ -321,7 +314,6 @@ impl<'tcx> CloneMap<'tcx> {
             prelude: IndexMap::new(),
             clone_graph: DiGraphMap::new(),
             dep_level: CloneLevel::Body,
-            used_types: Default::default(),
         }
     }
 
@@ -714,6 +706,7 @@ impl<'tcx> CloneMap<'tcx> {
         use petgraph::visit::Walker;
         let mut roots: IndexSet<_> = self.names.keys().cloned().collect();
 
+        let mut elab = CloneElaborator::new(self.param_env(ctx));
         // Update the clone graph with any new entries.
         self.update_graph(ctx, depth);
 
@@ -749,7 +742,9 @@ impl<'tcx> CloneMap<'tcx> {
                 continue;
             }
 
-            let Some(decl) = self.build_clone(ctx, node, depth) else { continue };
+            // HACK: REALLY BAD FOR PERFORAMNCE REMOVE ONCE `self.clone_graph` is removed (borrowck error)
+            let graph = self.clone_graph.clone();
+            let Some(decl) = elab.build_clone(ctx, &mut self, &graph, node, depth) else { continue };
             decls.push(decl);
         }
 
@@ -810,34 +805,6 @@ impl<'tcx> CloneMap<'tcx> {
         )
         .unwrap();
     }
-}
-
-pub(crate) fn base_subst<'tcx, N: Namer<'tcx>>(
-    ctx: &mut Why3Generator<'tcx>,
-    param_env: ParamEnv<'tcx>,
-    names: &mut N,
-    dep: DepNode<'tcx>,
-) -> Vec<CloneSubst> {
-    let (generics, substs) = if let DepNode::TyInv(ty, inv_kind) = dep {
-        let generics = inv_kind.generics(ctx.tcx);
-        let substs = inv_kind.tyinv_substs(ctx.tcx, ty);
-        (generics, substs)
-    } else if let Some((def_id, subst)) = dep.did() {
-        let generics = ty_param_names(ctx.tcx, def_id).collect();
-        (generics, subst)
-    } else {
-        return vec![];
-    };
-
-    let param_env = names.param_env(ctx);
-
-    let mut clone_substs = vec![];
-    for (name, arg) in iter::zip(generics, substs.types()) {
-        let ty = ctx.normalize_erasing_regions(param_env, arg);
-        let ty = backend::ty::translate_ty(ctx, names, rustc_span::DUMMY_SP, ty);
-        clone_substs.push(CloneSubst::Type(name.into(), ty));
-    }
-    clone_substs
 }
 
 /// Which level a clone appears at, determines its visibility in the different modules generated by Creusot
@@ -979,92 +946,4 @@ impl<'tcx, F: FnMut(Ty<'tcx>)> TypeVisitor<TyCtxt<'tcx>> for TyVisitor<'tcx, F> 
     }
 }
 
-struct CloneElaborator {}
-
-impl<'tcx> CloneElaborator {
-    fn build_clone(
-        &mut self,
-        ctx: &mut Why3Generator<'tcx>,
-        item: DepNode<'tcx>,
-        depth: CloneDepth,
-    ) -> Option<Decl> {
-        // Types can't be cloned, but are used (for now).
-        if let DepNode::Type(_) = item {
-            let (def_id, _) = item.did()?;
-            // check if type is not an assoc type
-            if util::item_type(ctx.tcx, def_id) == ItemType::Type {
-                let use_decl = self.used_types.insert(def_id).then(|| {
-                    if let Some(builtin) = get_builtin(ctx.tcx, def_id) {
-                        let name = QName::from_string(&builtin.as_str()).unwrap().module_qname();
-                        Use { name, as_: None, export: false }
-                    } else {
-                        let name = cloneable_name(ctx, item, CloneLevel::Body);
-                        Use { name: name.clone(), as_: Some(name), export: false }
-                    }
-                });
-                return use_decl.map(Decl::UseDecl);
-            }
-        }
-
-        let mut clone_subst = base_subst(ctx, self, item);
-        trace!("base substs of {item:?}: {clone_subst:?}");
-
-        let outbound: Vec<_> = self.clone_graph.neighbors_directed(item, Outgoing).collect();
-
-        let level_of_item = match (depth, self.names[&item].opaque) {
-            // We are requesting a deep clone of an opaque thing: stop at the contract
-            (CloneDepth::Deep, CloneOpacity::Opaque) => CloneLevel::Contract,
-            // Otherwise, go deep and get the body
-            (CloneDepth::Deep, _) => CloneLevel::Body,
-            // If we are only doing shallow clones, stop at the signature (no contracts)
-            (CloneDepth::Shallow, _) => CloneLevel::Signature,
-        };
-
-        // Grab definitions from all of our dependencies
-        for dep in outbound {
-            let (edge_level, syms) = &self.clone_graph[(item, dep)];
-
-            if *edge_level > level_of_item {
-                continue;
-            };
-            trace!("dependency={:?} of={:?} syms={:?}", dep, item, syms);
-
-            match dep {
-                DepNode::Type(ty) => {
-                    for (nm, sym) in syms.clone() {
-                        let ty_name = nm.qname_ident(sym.ident());
-                        let ty = backend::ty::translate_ty(ctx, self, DUMMY_SP, ty);
-                        clone_subst.push(CloneSubst::Type(ty_name, ty))
-                    }
-                }
-                _ => {
-                    for (nm, sym) in syms {
-                        let elem = sym.to_subst(*nm, self.names[&dep].kind);
-                        clone_subst.push(elem);
-                    }
-                }
-            }
-        }
-
-        let use_axioms = item.is_inv()
-            || item.did().is_some_and(|(def_id, _)| {
-                ctx.item(def_id).map(|i| i.has_axioms()).unwrap_or(false)
-            });
-
-        if use_axioms {
-            clone_subst.push(CloneSubst::Axiom(None))
-        }
-
-        trace!(
-            "emit clone node={item:?} name={:?} as={:?}",
-            cloneable_name(ctx, item, level_of_item),
-            self.names[&item].kind.clone()
-        );
-
-        Some(Decl::Clone(DeclClone {
-            name: cloneable_name(ctx, item, level_of_item),
-            subst: clone_subst,
-            kind: self.names[&item].kind.clone().into(),
-        }))
-    }
-}
+type DepGraph<'tcx> = DiGraphMap<DepNode<'tcx>, (CloneLevel, IndexSet<(Kind, SymbolKind)>)>;
